revamp: product's bulkaction in Bulkaction component
